나눗셈 소수점 이하 나오도록 수정

+++ Calculator.java
@@ -10,8 +10,9 @@ public class Calculator {
 	static int multiply(int a, int b) {
 		return a*b;
 	}
-	static double divide(double a, double b) {
-		return a /b;
+	static double divide(int a, int b) {
+		return a/(double)b;
 	}
 
 }
+
